tests/performance/test_processing_speed.py
"""
Performance tests for document processing speed.

Tests verify that performance targets from PRD are met:
- Processing speed: 10-50 documents/minute
- Startup time: < 3 seconds
- Memory usage: < 500MB for 500 documents
- UI responsiveness: 60 FPS during processing

Use pytest-benchmark for accurate measurements.
Run with: pytest -m performance --benchmark-only
"""

import pytest
import time
from pathlib import Path
from docx import Document

# TODO: Update imports when processors are implemented
# from src.processors.search_replace import process_document
# from src.processors.metadata import update_metadata


@pytest.mark.performance
@pytest.mark.slow
@pytest.mark.skip(reason="Processor not yet implemented")
def test_search_replace_processing_speed(benchmark, docx_factory, tmp_path):
    """
    Test search/replace processor meets speed target.

    Target: 10-50 documents/minute = 1.2-6 seconds per document
    This test should complete in < 6 seconds per document.
    """
    # Create test document
    doc_path = docx_factory(
        content="Test paragraph. " * 100,  # ~100 paragraphs
        filename="speed_test.docx"
    )

    # config = {"find": "Test", "replace": "Sample", "use_regex": False}
    #
    # # Benchmark processing
    # result = benchmark(process_document, str(doc_path), config)
    #
    # # Verify performance
    # assert result.success is True
    # # Benchmark automatically measures and reports timing
    pass


@pytest.mark.performance
@pytest.mark.skip(reason="Processor not yet implemented")
def test_batch_processing_10_documents(benchmark, multiple_docx):
    """
    Test batch processing speed for 10 documents.

    Target: 10-50 docs/min → 10 docs should take 12-60 seconds
    """
    # config = {"find": "test", "replace": "sample", "use_regex": False}
    #
    # def batch_process():
    #     results = []
    #     for doc_path in multiple_docx:
    #         result = process_document(str(doc_path), config)
    #         results.append(result)
    #     return results
    #
    # results = benchmark(batch_process)
    #
    # # Verify all succeeded
    # assert all(r.success for r in results)
    # assert len(results) == 10
    pass


@pytest.mark.performance
@pytest.mark.slow
@pytest.mark.skip(reason="Requires large document set")
def test_memory_usage_500_documents(docx_factory, tmp_path):
    """
    Test memory usage remains under 500MB for 500 documents.

    Target: < 500MB memory footprint
    """
    import psutil
    import os

    process = psutil.Process(os.getpid())

    # Create 500 test documents
    docs = []
    for i in range(500):
        doc_path = docx_factory(
            content=f"Document {i} content.\n" * 10,
            filename=f"mem_test_{i:03d}.docx"
        )
        docs.append(doc_path)

    # Measure baseline memory
    baseline_memory = process.memory_info().rss / 1024 / 1024  # MB

    # Process all documents
    # config = {"find": "Document", "replace": "File", "use_regex": False}
    # for doc_path in docs:
    #     process_document(str(doc_path), config)

    # Measure peak memory
    peak_memory = process.memory_info().rss / 1024 / 1024  # MB
    memory_used = peak_memory - baseline_memory

    # Verify under 500MB target
    assert memory_used < 500, f"Memory usage {memory_used:.2f}MB exceeds 500MB target"


@pytest.mark.performance
@pytest.mark.skip(reason="MainWindow not yet implemented")
def test_application_startup_time(qtbot, benchmark):
    """
    Test application startup time is under 3 seconds.

    Target: < 3 seconds from launch to ready
    """
    # from src.ui.main_window import MainWindow
    #
    # def startup():
    #     window = MainWindow()
    #     qtbot.addWidget(window)
    #     window.show()
    #     return window
    #
    # window = benchmark(startup)
    #
    # # Benchmark reports timing automatically
    # # Target: < 3000ms (3 seconds)
    # assert window.isVisible()
    pass


@pytest.mark.performance
@pytest.mark.skip(reason="Processor not yet implemented")
def test_large_document_processing(benchmark, docx_factory):
    """
    Test processing large document (50+ pages).

    Verifies performance scales reasonably with document size.
    """
    # Create large document (~50 pages)
    large_content = "\n".join([f"Paragraph {i} with content." for i in range(500)])
    doc_path = docx_factory(
        content=large_content,
        filename="large_doc.docx"
    )

    # config = {"find": "Paragraph", "replace": "Section", "use_regex": False}
    #
    # result = benchmark(process_document, str(doc_path), config)
    #
    # # Should complete in reasonable time
    # assert result.success is True
    # assert result.changes_made == 500
    pass


@pytest.mark.performance
@pytest.mark.skip(reason="Processor not yet implemented")
def test_regex_performance_vs_literal(benchmark, docx_factory):
    """
    Compare regex vs literal search performance.

    Literal search should be faster than regex for simple patterns.
    """
    doc_path = docx_factory(
        content="Test content. " * 1000,
        filename="perf_compare.docx"
    )

    # Test literal search
    # config_literal = {"find": "Test", "replace": "Sample", "use_regex": False}
    # start = time.perf_counter()
    # result_literal = process_document(str(doc_path), config_literal)
    # time_literal = time.perf_counter() - start

    # Reset document
    doc_path = docx_factory(
        content="Test content. " * 1000,
        filename="perf_compare2.docx"
    )

    # Test regex search
    # config_regex = {"find": "Test", "replace": "Sample", "use_regex": True}
    # start = time.perf_counter()
    # result_regex = process_document(str(doc_path), config_regex)
    # time_regex = time.perf_counter() - start

    # Literal should be faster
    # assert time_literal < time_regex
    # assert result_literal.changes_made == result_regex.changes_made
    pass


@pytest.mark.performance
def test_database_query_performance(test_db, benchmark):
    """
    Test database operations complete quickly.

    Target: < 100ms for typical queries
    """
    if test_db is None:
        pytest.skip("DatabaseManager not yet implemented")

    # def query_jobs():
    #     return test_db.get_all_jobs()
    #
    # # Insert test data
    # for i in range(100):
    #     test_db.create_job(operation="test", config={})
    #
    # # Benchmark query
    # results = benchmark(query_jobs)
    #
    # assert len(results) == 100
    # # Benchmark reports timing - should be << 100ms
    pass


@pytest.mark.performance
@pytest.mark.skip(reason="Worker not yet implemented")
def test_multiprocessing_pool_overhead(benchmark, multiple_docx):
    """
    Test multiprocessing pool initialization overhead.

    Verifies that pool creation doesn't dominate processing time.
    """
    # from multiprocessing import Pool
    # from src.processors.search_replace import process_document
    #
    # config = {"find": "test", "replace": "sample", "use_regex": False}
    #
    # def process_with_pool():
    #     with Pool(processes=4) as pool:
    #         results = pool.starmap(
    #             process_document,
    #             [(str(doc), config) for doc in multiple_docx]
    #         )
    #     return results
    #
    # results = benchmark(process_with_pool)
    #
    # # Pool overhead should be minimal compared to processing time
    # assert all(r.success for r in results)
    pass
